added log4j2 and logging by config  properties

+++ service/src/main/java/com/senla/library/service/impl/RoleServiceImpl.java
@@ -1,10 +1,14 @@
 package com.senla.library.service.impl;
 
 import com.senla.library.dto.RoleDTO;
+import com.senla.library.entity.Rent;
 import com.senla.library.entity.Role;
 import com.senla.library.mapper.RoleMapper;
 import com.senla.library.repository.RoleRepository;
 import com.senla.library.service.RoleService;
+import com.senla.library.service.exception.ResourceDuplicationException;
+import com.senla.library.service.exception.ResourceNotFoundException;
+import com.senla.library.service.exception.ResourceNotSave;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -30,28 +34,38 @@ public class RoleServiceImpl implements RoleService {
     
     @Override
     public RoleDTO save(RoleDTO dto) {
-        return mapper.toDto(repository.save(mapper.toEntity(dto)));
-    }
-    
-    @Override
-    public void save(List<RoleDTO> dtos) {
-        repository.saveAll(mapper.toEntityList(dtos));
+        Role role = mapper.toEntity(dto);
+        if(repository.existsByName(role.getName())){
+            throw new ResourceDuplicationException("CONFLICT, error saving data, " +
+                    "the database contains such data");
+        }else {
+            return mapper.toDto(repository.save(role));
+        }
     }
     
     @Override
     public void deleteById(Long id) {
-        repository.deleteById(id);
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+        }else {
+            throw new ResourceNotFoundException("Failed to delete by primary key ");
+        }
     }
     
     @Override
-    public Optional<RoleDTO> findById(Long id) {
-        Optional<Role> entityOptional = repository.findById(id);
-        return entityOptional.map(entity -> Optional.ofNullable(mapper.toDto(entity))).orElse(null);
+    public RoleDTO findById(Long id) {
+        return mapper.toDto(repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Role by id not found")));
     }
     
     @Override
     public List<RoleDTO> findAll() {
-        return mapper.toDtoList((List<Role>) repository.findAll());
+        Iterable<Role> all = repository.findAll();
+        List<Role>roles=(List<Role>)all;
+        if(roles.isEmpty() || roles==null){
+            throw new ResourceNotFoundException("Rents not found");
+        }else {
+            return mapper.toDtoList(roles);
+        }
     }
     
     @Override
@@ -63,10 +77,10 @@ public class RoleServiceImpl implements RoleService {
     
     @Override
     public RoleDTO updateById(RoleDTO dto) {
-        Optional<RoleDTO> optionalDto = findById(dto.getId());
-        if(optionalDto.isPresent()){
+        if(repository.existsById(dto.getId())){
             return save(dto);
+        } else {
+            throw new ResourceNotFoundException("update failed no record with this id");
         }
-        return null;
     }
 }
